+++ lib/rlm/config.rb
@@ -2,7 +2,7 @@
 
 module RLM
   class Config
-    attr_accessor :root_lm, :sub_lm, :sandbox, :cache, :default_limits, :trace_store
+    attr_accessor :root_lm, :sub_lm, :sandbox, :cache, :default_limits, :trace_store, :tool_authorizer
     attr_writer :logger
 
     def initialize
@@ -12,6 +12,7 @@ module RLM
       @cache = nil
       @default_limits = Limits.new
       @trace_store = nil
+      @tool_authorizer = nil
       @logger = nil
     end